split out metrics into catwalk.metrics
